Moved bintray plugins to specific package
