+++ src/main/groovy/org/shipkit/internal/gradle/ShipkitJavaPlugin.java
@@ -3,9 +3,9 @@ package org.shipkit.internal.gradle;
 import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
+import org.shipkit.internal.gradle.bintray.BintrayReleasePlugin;
 import org.shipkit.internal.gradle.java.JavaBintrayPlugin;
 import org.shipkit.internal.gradle.java.PomContributorsPlugin;
-import org.shipkit.internal.gradle.release.BintrayReleasePlugin;
 
 /**
  * Continuous delivery for Java with Travis and Bintray.
