+++ src/main/resources/META-INF/gradle-plugins/org.shipkit.bintray-release.properties
@@ -1,1 +1,1 @@
-implementation-class=org.shipkit.internal.gradle.release.BintrayReleasePlugin
+implementation-class=org.shipkit.internal.gradle.bintray.BintrayReleasePlugin
