+++ src/main/resources/META-INF/gradle-plugins/org.shipkit.bintray.properties
@@ -1,1 +1,1 @@
-implementation-class=org.shipkit.internal.gradle.ShipkitBintrayPlugin
+implementation-class=org.shipkit.internal.gradle.bintray.ShipkitBintrayPlugin
