+++ src/test/groovy/org/shipkit/internal/gradle/java/ComparePublicationsPluginTest.groovy
@@ -5,9 +5,9 @@ import org.gradle.testfixtures.ProjectBuilder
 import org.shipkit.gradle.ShipkitConfiguration
 import org.shipkit.gradle.java.ComparePublicationsTask
 import org.shipkit.gradle.java.DownloadPreviousPublicationsTask
-import org.shipkit.internal.gradle.ShipkitBintrayPlugin
-import org.shipkit.internal.gradle.version.VersioningPlugin
+import org.shipkit.internal.gradle.bintray.ShipkitBintrayPlugin
 import org.shipkit.internal.gradle.configuration.ShipkitConfigurationPlugin
+import org.shipkit.internal.gradle.version.VersioningPlugin
 import testutil.PluginSpecification
 
 class ComparePublicationsPluginTest extends PluginSpecification {
